+++ CHANGES.rst
@@ -22,6 +22,7 @@
 * Replace ``csrf_form`` in template context with ``csrf_token``, regardless of whether
   global CSRF is enabled
 * Optional support for Flask-DebugToolbar and LineProfiler extension
+* SelectField now supports optgroup grouping
 
 0.3.0
 -----